requirements.txt
opencv-python
numpy
dlib
face-recognition
face-recognition-models
setuptools

crewai>=0.70.0
crewai-tools>=0.12.0
langchain-groq>=0.1.0
langchain-ollama>=0.1.0
python-dotenv>=1.0.0
PyPDF2>=3.0.0
pypdf>=4.0.0
deepface>=0.0.90
tf-keras>=2.16.0
mediapipe>=0.10.0
ultralytics>=8.0.0
pydantic>=2.0.0

